Add authentication and authorization based on roles and credentials, jwt based. (#1)
Features implemented:

Authentication and Authorization based on username and password
Security based on JWT (Bearer)
Implemented Refresh Token
Implemented Access Token after registration/login
Everything organized in a hexagonal-based architecture.
Updated to Java 21 and lastest spring boot version.
Updated to gradle 8.6
Move to postgresql database
For the moment tokens are stored in database. A good practice is to create a cache (redis preferably) in order to mark tokens as 'revoked/expired' when someone logout. Storing tokens in database IS NOT a good practice and not recomended, but for the moment that could work.